+++ Changelog.md
@@ -10,6 +10,7 @@ Compiler Features:
  * SMTChecker: Output values for ``block.*``, ``msg.*`` and ``tx.*`` variables that are present in the called functions.
  * Standard JSON: Accept nested brackets in step sequences passed to ``settings.optimizer.details.yulDetails.optimizerSteps``.
  * Standard JSON: Add ``settings.debug.debugInfo`` option for selecting how much extra debug information should be included in the produced EVM assembly and Yul code.
+ * TypeChecker: Improve error message in implicit function type conversions failures from and to builtin function types.
 
 
 Bugfixes: